pyproject.toml
[tool.poetry]
name = "pga-graphiti"
version = "0.1.0"
description = ""
authors = ["katherineforbes <katanneforbes@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.13,<4.0"
graphiti-core = "^0.10.5"
pygithub = "^2.6.1"
python-dotenv = "^1.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
